segmentation/requirements.txt
Django==3.0.8
imageio==2.8.0
matplotlib==3.1.1
numpy==1.17.1
opencv-python==4.2.0.32
pandas==1.0.4
Pillow==6.2.0
python-dateutil==2.8.0
requests==2.24.0
scikit-learn==0.22.1
scipy==1.2.3
tqdm==4.46.0
gunicorn==20.0.4